+++ CHANGES.md
@@ -5,7 +5,7 @@
 * `bayes.update` now supports `lognorm` and `gamma` distributions in addition to `norm` and `beta`.
 * Allow `zero_inflated` to accept 0 or 1 as probabilities.
 * **[New feature]** Added `die(sides, explode_on=None)` and `coin()` as distribution objects for dice and coins. These support all distribution operations like `~`, `@`, `+`, `-`, etc. The `die` distribution also supports "exploding dice" mechanics via the `explode_on` parameter.
-* **[Deprecation]** `roll_die` and `flip_coin` functions are now deprecated. Use `sq.die(sides) @ n` or `~sq.coin()` instead.
+* **[Breaking change]** Removed `roll_die` and `flip_coin` functions. Use `sq.die(sides) @ n` or `~sq.coin()` instead.
 
 ## v0.29 - latest release